+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2021-05-2
+
+### Removed
+
+-   Disabled preloading till I found a solution.
+
 ### Fixed
 
 -   Support for Pinecone Router 0.1.1
@@ -25,4 +31,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [unreleased]: https://github.com/pinecone-router/middleware-render/compare/0.0.0...HEAD
 [0.0.0]: https://github.com/pinecone-router/middleware-render/compare/0.0.0...0.0.0
-[0.0.1]: https://github.com/pinecone-router/middleware-render/compare/0.0.0...v0.0.1
+[0.0.1]: https://github.com/pinecone-router/middleware-render/compare/0.0.0...0.0.1
+[0.0.1]: https://github.com/pinecone-router/middleware-render/compare/0.0.1...0.0.2